Improvements to dynamic model page
