+++ entity-framework/core/modeling/dynamic-model.md
@@ -1,20 +1,28 @@
 ---
 title: Alternating between multiple models with the same DbContext type - EF Core
 author: AndriySvyryd
-ms.date: 12/10/2017
+ms.date: 01/03/2020
 ms.assetid: 3154BF3C-1749-4C60-8D51-AE86773AA116
 uid: core/modeling/dynamic-model
 ---
 # Alternating between multiple models with the same DbContext type
 
-The model built in `OnModelCreating` could use a property on the context to change how the model is built. For example it could be used to exclude a certain property:
+The model built in `OnModelCreating` can use a property on the context to change how the model is built. For example, suppose you wanted to configure an entity differently based on some property:
 
-[!code-csharp[Main](../../../samples/core/DynamicModel/DynamicContext.cs?name=Class)]
+[!code-csharp[Main](../../../samples/core/Modeling/DynamicModel/DynamicContext.cs?name=OnModelCreating)]
+
+Unfortunately, this code wouldn't work as-is, since EF builds the model and runs `OnModelCreating` only once, caching the result for performance reasons. However, you can hook into the model caching mechanism to make EF aware of the property producing different models.
 
 ## IModelCacheKeyFactory
 
-However if you tried doing the above without additional changes you would get the same model every time a new context is created for any value of `IgnoreIntProperty`. This is caused by the model caching mechanism EF uses to improve the performance by only invoking `OnModelCreating` once and caching the model.
+EF uses the `IModelCacheKeyFactory` to generate cache keys for models; by default, EF assumes that for any given context type the model will be the same, so the default implementation of this service returns a key that just contains the context type. To produce different models from the same context type, you need to replace the `IModelCacheKeyFactory` service with the correct  implementation; the generated key will be compared to other model keys using the `Equals` method, taking into account all the variables that affect the model:
 
-By default EF assumes that for any given context type the model will be the same. To accomplish this the default implementation of `IModelCacheKeyFactory` returns a key that just contains the context type. To change this you need to replace the `IModelCacheKeyFactory` service. The new implementation needs to return an object that can be compared to other model keys using the `Equals` method that takes into account all the variables that affect the model:
+The following implementation takes the `IgnoreIntProperty` into account when producing a model cache key:
 
-[!code-csharp[Main](../../../samples/core/DynamicModel/DynamicModelCacheKeyFactory.cs?name=Class)]
+[!code-csharp[Main](../../../samples/core/Modeling/DynamicModel/DynamicModelCacheKeyFactory.cs?name=DynamicModel)]
+
+Finally, register your new `IModelCacheKeyFactory` in your context's `OnConfiguring`:
+
+[!code-csharp[Main](../../../samples/core/Modeling/DynamicModel/DynamicContext.cs?name=OnConfiguring)]
+
+See the [full sample project](https://github.com/aspnet/EntityFramework.Docs/tree/master/samples/core/Modeling/DynamicModel) for more context.
