+++ samples/core/Modeling/DynamicModel/ConfigurableEntity.cs
