+++ samples/core/Modeling/DynamicModel/DynamicModel.csproj
